Move fixLocalMediaURLs to WordPressData
